fix: removed seed because blocks on homepage changed and IDK how to seed richText fields
fix: added all blocks to Posts collections as placeholders so they are saved in the db and I can generate types from them. Otherwise, the code breaks.

fix: renamed FAQ section's slug so it can re-generate broken tables

WARNING: when multiple prompts are present, the hook will execute max 1, and move on to the next command